pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "jobapply"
version = "0.1.0"
description = "Automated job application agent for robotics and ADAS engineers"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.25.0",
    "playwright>=1.44.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "python-docx>=1.1.0",
    "weasyprint>=62.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
    "PyPDF2>=3.0.0",
    "rapidfuzz>=3.9.0",
]

[project.scripts]
jobapply = "main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "core*", "ats_adapters*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
